requirements.txt
psutil==5.9.7
Pillow==10.2.0
